Hack: avoid protocol change check
* https://github.com/webpack/webpack-dev-server/issues/863

WDS changes the entry protocol based on the serving page to avoid
security exceptions. This doesn't work in our case because WDS is
fronted by a HTTPS web server, but WDS itself is running HTTP. A better
solution for our setup right now is to avoid the check and accept a
security exception in dev.

The issue outlines a suggested fix which I will implement if there is
interest.